More refactor

+++ Hid.Net/DeviceInformation.cs
@@ -3,8 +3,8 @@
 
     public class DeviceInformation
     {
-        public ushort VendorId { get; set; }
-        public ushort ProductId { get; set; }
+        public int VendorId { get; set; }
+        public int ProductId { get; set; }
         public string DeviceId { get; set; }
     }
 }
